drivers: tt_timer: move timer code into tenstorrent driver
Move code for ARC REFCLK into a tenstorrent-specific driver
