+++ app/smc/prj.conf
@@ -34,3 +34,7 @@ CONFIG_TT_BH_ARC_SYSINIT=y
 
 # Enable required libraries
 CONFIG_TT_FW_TABLE=y
+
+# Enable required drivers
+CONFIG_TT_TIMER=y
+CONFIG_TT_BH_ARC_TIMER=y
