+++ drivers/Kconfig
@@ -11,6 +11,7 @@ rsource "reset/Kconfig"
 rsource "sensor/Kconfig"
 rsource "serial/Kconfig"
 rsource "smbus/Kconfig"
+rsource "tenstorrent/Kconfig"
 # zephyr-keep-sorted-stop
 
 endmenu
