+++ drivers/tenstorrent/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Copyright (c) 2025 Tenstorrent AI ULC
+# SPDX-License-Identifier: Apache-2.0
+
+# zephyr-keep-sorted-start
+add_subdirectory_ifdef(CONFIG_TT_TIMER timer)
+# zephyr-keep-sorted-stop
