+++ drivers/tenstorrent/timer/blackhole/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (c) 2025 Tenstorrent AI ULC
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_sources(timer.c)
