+++ lib/tenstorrent/bh_arc/clock_wave.c
@@ -10,7 +10,7 @@
 #include <tenstorrent/msg_type.h>
 #include <tenstorrent/msgqueue.h>
 
-#include "timer.h"
+#include <tenstorrent/drivers/timer/timer.h>
 #include <tenstorrent/reg.h>
 
 #define PLL_CNTL_WRAPPER_CLOCK_WAVE_CNTL_REG_ADDR 0x80020038
