+++ lib/tenstorrent/bh_arc/spi_controller.c
@@ -11,7 +11,7 @@
 
 #include "pll.h"
 #include <tenstorrent/reg.h>
-#include "timer.h"
+#include <tenstorrent/drivers/timer/timer.h>
 
 #define SPICLK_FREQ_MHZ 40
